restricting pages and registration complete

+++ base/views.py
@@ -1,15 +1,23 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
+from django.http import HttpResponse
 
 # dont name it login as it will conflict with django login function
 def loginPage(request):
+    page = 'login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         
         try:
@@ -25,12 +33,32 @@ def loginPage(request):
         else: 
             messages.error(request, 'Username or password does not exist.')
         
-    context = {}
+    context = {'page': page}
     return render(request, "base/login_register.html", context)
+
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False) # commit = false means dont save yet for now because we want to modify it
+            user.username = user.username.lower() # make username lowercase
+            user.save() # now save
+            login(request, user) 
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration.')
+        
+    context = {'form': form}
+    return render(request, "base/login_register.html", context)
+
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -56,8 +84,14 @@ def room(request, pk):
     context = {'room': room}
     return render(request=request, template_name="base/room.html", context=context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
+    
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+    
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -67,9 +101,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, "base/room_form.html", context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)  # pre-fill with existing data
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -79,8 +118,13 @@ def updateRoom(request, pk):
 
     return render(request, "base/room_form.html", {'form': form})
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
 
     if request.method == 'POST':
         room.delete()
